[FEATURE] Add backend module with schema inspector
This adds a backend module which can be enabled in EXT:schemaker extension configuration and renders ViewHelper documentation similar to the frontend plugin which also renders the official FluidTYPO3 ViewHelper documentation, but renders based on the local classes (including any delivered by third-party extensions). The module integrates with the user session, saving the location for a better experience when switching between backend modules.

Note that for third party extension support it is necessary to configure an extension key to PHP namespace prefix map in `module.tx_schemaker.settings.extensionKeyClassNamespaceMap` with an array like `vhs = FluidTYPO3.Vhs` etc.

The change also:

* Removes the history component
* Removes the automatically rendered tag examples
* Adds a warning also in FE if no "markdown" utility is detected/configured
* Displays the raw markdown source using `nl2br` as fallback when no "markdown" is found